+++ packager/pincab/pinballX.py
@@ -33,21 +33,45 @@ class PinballX:
         self.logger.info("* Pinball X files")
         for file in Path(self.pinballX_path).glob('**/*%s*' % table_name):
             if "Flyer Images\\Back" in str(file.parent):
-                package.add_file(file, 'media/Flyers Back', dstFile=file.stem + ".back" + file.suffix)
+                if is_suffix(file, '.back'):
+                    package.add_file(file, 'media/Flyers Back', dst_file=file.stem + file.suffix)
+                else:
+                    package.add_file(file, 'media/Flyers Back', dst_file=file.stem + ".back" + file.suffix)
             elif "Flyer Images\\Front" in str(file.parent):
-                package.add_file(file, 'media/Flyers Front', dstFile=file.stem + ".front" + file.suffix)
+                if is_suffix(file, '.front'):
+                    package.add_file(file, 'media/Flyers Front', dst_file=file.stem + file.suffix)
+                else:
+                    package.add_file(file, 'media/Flyers Front', dst_file=file.stem + ".front" + file.suffix)
             elif "Flyer Images\\Inside1" in str(file.parent):
-                package.add_file(file, 'media/Flyers Inside', dstFile=file.stem + ".inside1" + file.suffix)
+                if is_suffix(file, '.inside1'):
+                    package.add_file(file, 'media/Flyers Inside', dst_file=file.stem + file.suffix)
+                else:
+                    package.add_file(file, 'media/Flyers Inside', dst_file=file.stem + ".inside1" + file.suffix)
             elif "Flyer Images\\Inside2" in str(file.parent):
-                package.add_file(file, 'media/Flyers Inside', dstFile=file.stem + ".inside2" + file.suffix)
+                if is_suffix(file, '.inside2'):
+                    package.add_file(file, 'media/Flyers Inside', dst_file=file.stem + file.suffix)
+                else:
+                    package.add_file(file, 'media/Flyers Inside', dst_file=file.stem + ".inside2" + file.suffix)
             elif "Flyer Images\\Inside3" in str(file.parent):
-                package.add_file(file, 'media/Flyers Inside', dstFile=file.stem + ".inside3" + file.suffix)
+                if is_suffix(file, '.inside3'):
+                    package.add_file(file, 'media/Flyers Inside', dst_file=file.stem + file.suffix)
+                else:
+                    package.add_file(file, 'media/Flyers Inside', dst_file=file.stem + ".inside3" + file.suffix)
             elif "Flyer Images\\Inside4" in str(file.parent):
-                package.add_file(file, 'media/Flyers Inside', dstFile=file.stem + ".inside4" + file.suffix)
+                if is_suffix(file, '.inside4'):
+                    package.add_file(file, 'media/Flyers Inside', dst_file=file.stem + file.suffix)
+                else:
+                    package.add_file(file, 'media/Flyers Inside', dst_file=file.stem + ".inside4" + file.suffix)
             elif "Flyer Images\\Inside5" in str(file.parent):
-                package.add_file(file, 'media/Flyers Inside', dstFile=file.stem + ".inside5" + file.suffix)
+                if is_suffix(file, '.inside5'):
+                    package.add_file(file, 'media/Flyers Inside', dst_file=file.stem + file.suffix)
+                else:
+                    package.add_file(file, 'media/Flyers Inside', dst_file=file.stem + ".inside5" + file.suffix)
             elif "Flyer Images\\Inside6" in str(file.parent):
-                package.add_file(file, 'media/Flyers Inside', dstFile=file.stem + ".inside6" + file.suffix)
+                if is_suffix(file, '.inside6'):
+                    package.add_file(file, 'media/Flyers Inside', dst_file=file.stem + file.suffix)
+                else:
+                    package.add_file(file, 'media/Flyers Inside', dst_file=file.stem + ".inside6" + file.suffix)
             elif "High Scores\\Visual Pinball" in str(file.parent):
                 package.add_file(file, 'media/HighScores')
             elif "Instruction Cards" in str(file.parent):